irpc

Attribute Macro rpc_requests

#[rpc_requests]

Processes an RPC request enum and generates channel implementations.

This macro takes a protocol enum where each variant represents a different RPC request type and generates the necessary channel implementations for each request.

Macro Arguments

First positional argument (required): The service type that will handle these requests

message (optional): Generate an extended enum wrapping each type in WithChannels<T, Service>

alias (optional): Generate type aliases with the given suffix for each WithChannels<T, Service>

Variant Attributes

Individual enum variants can be annotated with the #[rpc(...)] attribute to specify channel types:

#[rpc(tx=SomeType)]: Specify the transmitter/sender channel type (required)

#[rpc(tx=SomeType, rx=OtherType)]: Also specify a receiver channel type (optional)

If rx is not specified, it defaults to NoReceiver.

Examples

Basic usage:

#[rpc_requests(ComputeService)]
enum ComputeProtocol {
    #[rpc(tx=oneshot::Sender<u128>)]
    Sqr(Sqr),
    #[rpc(tx=oneshot::Sender<i64>)]
    Sum(Sum),
}

With a message enum:

#[rpc_requests(ComputeService, message = ComputeMessage)]
enum ComputeProtocol {
    #[rpc(tx=oneshot::Sender<u128>)]
    Sqr(Sqr),
    #[rpc(tx=oneshot::Sender<i64>)]
    Sum(Sum),
}

With type aliases:

#[rpc_requests(ComputeService, alias = "Msg")]
enum ComputeProtocol {
    #[rpc(tx=oneshot::Sender<u128>)]
    Sqr(Sqr), // Generates type SqrMsg = WithChannels<Sqr, ComputeService>
    #[rpc(tx=oneshot::Sender<i64>)]
    Sum(Sum), // Generates type SumMsg = WithChannels<Sum, ComputeService>
}