irpc::channel::mpsc

Function channel

pub fn channel<T>(buffer: usize) -> (Sender<T>, Receiver<T>)

Create a local mpsc sender and receiver pair, with the given buffer size.

This is currently using a tokio channel pair internally.