irpc::channel::mpsc

Trait DynSender

pub trait DynSender<T>:
    Debug
    + Send
    + Sync
    + 'static {
    // Required methods
    fn send(
        &self,
        value: T,
    ) -> Pin<Box<dyn Future<Output = Result<(), SendError>> + Send + '_>>;
    fn try_send(
        &self,
        value: T,
    ) -> Pin<Box<dyn Future<Output = Result<bool, SendError>> + Send + '_>>;
    fn closed(&self) -> Pin<Box<dyn Future<Output = ()> + Send + Sync + '_>>;
    fn is_rpc(&self) -> bool;
}

A sender that can be wrapped in a Arc<dyn DynSender<T>>.

Required Methods

fn send( &self, value: T, ) -> Pin<Box<dyn Future<Output = Result<(), SendError>> + Send + '_>>

Send a message.

For the remote case, if the message can not be completely sent, this must return an error and disable the channel.

fn try_send( &self, value: T, ) -> Pin<Box<dyn Future<Output = Result<bool, SendError>> + Send + '_>>

Try to send a message, returning as fast as possible if sending is not currently possible.

For the remote case, it must be guaranteed that the message is either completely sent or not at all.

fn closed(&self) -> Pin<Box<dyn Future<Output = ()> + Send + Sync + '_>>

Await the sender close

fn is_rpc(&self) -> bool

True if this is a remote sender