irpc::channel::mpsc

Enum Sender

pub enum Sender<T> {
    Tokio(Sender<T>),
    Boxed(Arc<dyn DynSender<T>>),
}

Single producer, single consumer sender.

For the local case, this wraps a tokio::sync::mpsc::Sender.

Variants

Tokio(Sender<T>)

Boxed(Arc<dyn DynSender<T>>)

Implementations

impl<T> Sender<T>

pub fn is_rpc(&self) -> bool
where T: 'static,

pub async fn closed(&self)
where T: RpcMessage,

pub fn into_sink(self) -> impl Sink<T, Error = SendError> + Send + 'static
where T: RpcMessage,

impl<T: RpcMessage> Sender<T>

pub async fn send(&self, value: T) -> Result<(), SendError>

Send a message and yield until either it is sent or an error occurs.

Cancellation safety

If the future is dropped before completion, and if this is a remote sender, then the sender will be closed and further sends will return an [io::Error] with [io::ErrorKind::BrokenPipe]. Therefore, make sure to always poll the future until completion if you want to reuse the sender or any clone afterwards.

pub async fn try_send(&mut self, value: T) -> Result<bool, SendError>

Try to send a message, returning as fast as possible if sending is not currently possible. This can be used to send ephemeral messages.

For the local case, this will immediately return false if the channel is full.

For the remote case, it will attempt to send the message and return false if sending the first byte fails, otherwise yield until the message is completely sent or an error occurs. This guarantees that the message is sent either completely or not at all.

Returns true if the message was sent.

Cancellation safety

If the future is dropped before completion, and if this is a remote sender, then the sender will be closed and further sends will return an [io::Error] with [io::ErrorKind::BrokenPipe]. Therefore, make sure to always poll the future until completion if you want to reuse the sender or any clone afterwards.